Extract Device class from Application class.
