--- Game/src/THEngine/Common/THFrameTimer.cpp
@@ -1,20 +1,0 @@
-#include "THFrameTimer.h"
-
-namespace THEngine
-{
-	FrameTimer::FrameTimer()
-	{
-		frameLeft = 0;
-		paused = false;
-	}
-
-	FrameTimer::~FrameTimer()
-	{
-
-	}
-
-	Object* FrameTimer::Clone()
-	{
-		return new FrameTimer(*this);
-	}
-}
