+++ Game/src/THEngine/Core/3D/THMesh.cpp
@@ -1,6 +1,6 @@
 #include "THMesh.h"
 #include <Core\THGame.h>
-#include <Platform\THApplication.h>
+#include <Platform\THDevice.h>
 #include <Asset\THAssetManager.h>
 #include <Asset\THTexture.h>
 #include <Renderer\THRenderPipeline.h>
@@ -10,7 +10,6 @@ namespace THEngine
 {
 	Mesh::Mesh()
 	{
-
 	}
 
 	Mesh::Mesh(const Mesh& mesh) : GameObject(mesh)
@@ -52,7 +51,7 @@ namespace THEngine
 
 	void Mesh::InitVertexBuffer(int size)
 	{
-		auto device = Application::GetInstance()->GetDevice();
+		auto device = Device::GetInstance()->GetD3DDevice();
 
 		device->CreateVertexBuffer(size * sizeof(MeshVertex), 0,
 			D3DFVF_XYZ | D3DFVF_NORMAL | D3DFVF_TEX1, D3DPOOL_MANAGED, &vertexBuffer, nullptr);
@@ -82,7 +81,7 @@ namespace THEngine
 
 	void Mesh::DrawGeometry()
 	{
-		auto app = Application::GetInstance();
+		auto device = Device::GetInstance();
 
 		if (this->mesh)
 		{
@@ -92,16 +91,16 @@ namespace THEngine
 		{
 			D3DVERTEXBUFFER_DESC vbdesc;
 			this->vertexBuffer->GetDesc(&vbdesc);
-			app->GetDevice()->SetFVF(vbdesc.FVF);
+			device->GetD3DDevice()->SetFVF(vbdesc.FVF);
 
 			switch (this->primitiveType)
 			{
 			case Mesh::TRIANGLE_LIST:
-				app->GetDevice()->DrawPrimitive(D3DPT_TRIANGLELIST, 0, this->vertexCount / 3);
+				device->GetD3DDevice()->DrawPrimitive(D3DPT_TRIANGLELIST, 0, this->vertexCount / 3);
 				break;
 			case Mesh::TRIANGLE_STRIP:
-				app->GetDevice()->SetStreamSource(0, this->vertexBuffer, 0, sizeof(MeshVertex));
-				app->GetDevice()->DrawPrimitive(D3DPT_TRIANGLESTRIP, 0, this->vertexCount - 2);
+				device->GetD3DDevice()->SetStreamSource(0, this->vertexBuffer, 0, sizeof(MeshVertex));
+				device->GetD3DDevice()->DrawPrimitive(D3DPT_TRIANGLESTRIP, 0, this->vertexCount - 2);
 				break;
 			}
 		}
@@ -118,7 +117,7 @@ namespace THEngine
 		{
 			if (mesh->LoadFromX(filePath) == false)
 			{
-				Exception* exception = new Exception((String)"加载模型文件失败：" + filePath + "原因是:\n" 
+				Exception* exception = new Exception((String)"加载模型文件失败：" + filePath + "原因是:\n"
 					+ exceptionManager->GetException()->GetInfo());
 				exceptionManager->PushException(exception);
 				delete mesh;
@@ -154,12 +153,11 @@ namespace THEngine
 	//////////////////////////////////////////////////////////////////////
 	Mesh::D3DMesh::D3DMesh()
 	{
-	
 	}
 
 	Mesh::D3DMesh::D3DMesh(const D3DMesh& d3dMesh) : Object(d3dMesh)
 	{
-		auto device = Application::GetInstance()->GetDevice();
+		auto device = Device::GetInstance()->GetD3DDevice();
 
 		IDirect3DVertexBuffer9* vb;
 		D3DVERTEXBUFFER_DESC vbdesc;
@@ -197,14 +195,14 @@ namespace THEngine
 	{
 		D3DMesh* d3dMesh = new D3DMesh();
 
-		auto device = Application::GetInstance()->GetDevice();
+		auto device = Device::GetInstance()->GetD3DDevice();
 		auto exceptionManager = ExceptionManager::GetInstance();
 
 		ID3DXBuffer* adjacency = nullptr;
 		ID3DXBuffer* materialBuffer = nullptr;
 		DWORD numMaterials;
 
-		if (FAILED(D3DXLoadMeshFromX(filePath.GetBuffer(), D3DXMESH_MANAGED, device, &adjacency, &materialBuffer ,
+		if (FAILED(D3DXLoadMeshFromX(filePath.GetBuffer(), D3DXMESH_MANAGED, device, &adjacency, &materialBuffer,
 			nullptr, &numMaterials, &d3dMesh->mesh)))
 		{
 			Exception* exception = new Exception("无法加载模型文件。文件可能已损坏。");
@@ -242,7 +240,7 @@ namespace THEngine
 					return nullptr;
 				}
 				d3dMesh->materialList[i].texture->Retain();
-			}		
+			}
 			d3dMesh->materialList[i].ambient[0] = materials[i].MatD3D.Diffuse.r;
 			d3dMesh->materialList[i].ambient[1] = materials[i].MatD3D.Diffuse.g;
 			d3dMesh->materialList[i].ambient[2] = materials[i].MatD3D.Diffuse.b;
@@ -272,10 +270,9 @@ namespace THEngine
 
 	void Mesh::D3DMesh::DrawGeometry()
 	{
-		auto meshShader = Application::GetInstance()->GetRenderState()->GetCurrentShader();
+		auto meshShader = Device::GetInstance()->GetRenderState()->GetCurrentShader();
 		for (int i = 0; i < this->numMaterials; i++)
 		{
-			
 			Material& mat = this->materialList[i];
 			if (mat.texture)
 			{
