+++ Game/src/THEngine/Core/THCamera.cpp
@@ -4,7 +4,7 @@
 #include <Asset\THAssetManager.h>
 #include <Asset\THCubeMap.h>
 #include <Asset\THRenderTexture.h>
-#include <Platform\THApplication.h>
+#include <Platform\THDevice.h>
 #include <Renderer\THRenderPipeline.h>
 #include <Renderer\THSkyBoxRenderer.h>
 
@@ -12,12 +12,10 @@ namespace THEngine
 {
 	Camera::Camera()
 	{
-
 	}
 
 	Camera::Camera(const String& name) : name(name)
 	{
-
 	}
 
 	Camera::~Camera()
@@ -32,9 +30,8 @@ namespace THEngine
 
 	void Camera::Draw()
 	{
-
 	}
-	
+
 	void Camera::SetRenderTexture(RenderTexture* renderTexture)
 	{
 		TH_SET(this->renderTexture, renderTexture);
@@ -42,14 +39,14 @@ namespace THEngine
 
 	void Camera::SetupViewport(Layer* layer)
 	{
-		auto app = Application::GetInstance();
-		app->SetViewport(layer->GetLeft() + this->viewport.left, layer->GetTop() + this->viewport.top,
+		auto device = Device::GetInstance();
+		device->SetViewport(layer->GetLeft() + this->viewport.left, layer->GetTop() + this->viewport.top,
 			this->viewport.Width(), this->viewport.Height());
 	}
 
 	void Camera::Render(Layer* layer)
 	{
-		auto renderState = Application::GetInstance()->GetRenderState();
+		auto renderState = Device::GetInstance()->GetRenderState();
 		TH_SET(renderState->camera, this);
 	}
 
@@ -66,51 +63,47 @@ namespace THEngine
 
 	Camera2D::~Camera2D()
 	{
-
 	}
 
 	void Camera2D::Render(Layer* layer)
 	{
 		Camera::Render(layer);
 
-		auto app = Application::GetInstance();
+		auto device = Device::GetInstance();
 
 		SetupViewport(layer);
 
 		Vector3f pos = GetPosition();
-		app->SetOrtho(pos.x - GetWidth() / 2, pos.y - GetHeight() / 2, 
+		device->SetOrtho(pos.x - GetWidth() / 2, pos.y - GetHeight() / 2,
 			GetWidth(), GetHeight(), this->znear, this->zfar);
 
 		Matrix matrix;
 		Matrix::Identity(&matrix);
-		app->SetViewMatrix(matrix);
+		device->SetViewMatrix(matrix);
 
 		Game::GetInstance()->GetRenderPipeline()->Render();
 
-		Application::GetInstance()->ClearDepthBuffer();
+		Device::GetInstance()->ClearDepthBuffer();
 	}
 
 	/////////////////////////////////////////
 	Camera3D::Camera3D()
 	{
-
 	}
 
 	Camera3D::Camera3D(const String& name) : Camera(name)
 	{
-
 	}
 
 	Camera3D::~Camera3D()
 	{
-
 	}
 
 	void Camera3D::Render(Layer* layer)
 	{
 		Camera::Render(layer);
 
-		auto app = Application::GetInstance();
+		auto device = Device::GetInstance();
 
 		SetupViewport(layer);
 
@@ -123,16 +116,15 @@ namespace THEngine
 		Matrix::Perspective(&matrix, this->fov, (float)this->viewport.Width() / this->viewport.Height(),
 			this->znear, this->zfar);
 
-		app->SetProjectionMatrix(matrix);
+		device->SetProjectionMatrix(matrix);
 
-		Matrix::LookAt(&matrix, Vector3f(this->position.x, this->position.y,  this->position.z),
+		Matrix::LookAt(&matrix, Vector3f(this->position.x, this->position.y, this->position.z),
 			Vector3f(this->lookAt.x, this->lookAt.y, this->lookAt.z),
 			Vector3f(this->up.x, this->up.y, this->up.z));
-		app->SetViewMatrix(matrix);
+		device->SetViewMatrix(matrix);
 
 		Game::GetInstance()->GetRenderPipeline()->Render();
 
-		Application::GetInstance()->ClearDepthBuffer();
+		device->ClearDepthBuffer();
 	}
 }
-
