+++ Game/src/THEngine/Renderer/THRenderer.cpp
@@ -2,23 +2,23 @@
 #define TH_RENDERER_H
 
 #include "THRenderer.h"
-#include <Platform\THApplication.h>
+#include <Platform\THDevice.h>
 #include <Core\THGameObject.h>
 
 namespace THEngine
 {
 	void Renderer::SetupRenderState(GameObject* obj)
 	{
-		auto app = Application::GetInstance();
+		auto device = Device::GetInstance();
 
-		app->SetBlendMode(obj->GetBlendMode());
+		device->SetBlendMode(obj->GetBlendMode());
 	}
 
 	void Renderer::SetupWorldTransform(GameObject* object)
 	{
 		Matrix world;
 		object->GetWorldMatrix(&world);
-		Application::GetInstance()->SetWorldMatrix(world);
+		Device::GetInstance()->SetWorldMatrix(world);
 	}
 }
